pyproject.toml
[project]
name = "hartufo"
authors = [{name = "Johan Pauwels", email = "johan.pauwels@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: Apache Software License"]
dynamic = ["version", "description"]
requires-python = ">=3.7"
dependencies = [
    "matplotlib >=3.5",
    "netCDF4 >=1.5",
    "numpy >=1.22",
    "openpyxl >=3.1",
    "Pillow >=9",
    "pymatreader >=0.0.32",
    "samplerate >=0.1.0",
    "scipy >=1.7",
    "torchvision >=0.11",
]

[project.urls]
Home = "https://github.com/jpauwels/hartufo"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.sdist]
exclude = [".gitignore", "docs/", "tests/"]
